@@ -1,8 +1,9 @@
-var App = require('app');
-
-App.Step = Ember.Object.extend({
+var Step = Em.Object.extend({
   active: false,
   velocity: 0,
+  setup: function() {
+    this.get('lastUpdated');
+  }.on('init'),
   enabled: function() {
     return this.get('velocity') !== 0;
   }.property('velocity'),
@@ -14,13 +15,14 @@ App.Step = Ember.Object.extend({
       velocity: this.get('velocity')
     };
   }
-
 }).reopenClass({
   deserialiseArray: function(data) {
     return Em.A(
       data.map(function(item) {
-        return App.Step.create(item);
+        return Step.create(item);
       })
     );
   }
 });
+
+export default Step;
